Added tabs for Agency page
