+++ src/pages/counterparties/counterparty-navigation.js
@@ -5,7 +5,7 @@ import { NavLink } from 'react-router-dom';
 const CounterpartyNavigation = ({ match }) => {
     const { id } = match.params
     return (
-        <section className="counterparty-tabs-navigation">
+        <section style={{ height: '34px' }} className="counterparty-tabs-navigation">
             <NavLink to={`/counterparties/${id}/transfer-balance`}>
                 Balance
             </NavLink>
